requirements.txt
langchain-google-genai
opencv-python
langchain-core 
python-dotenv
fastapi 
uvicorn 
google-generativeai
python-multipart